Fixed IllegalStateException crash when in UpholdAccountActivity.onResume
